restructure main & app
